Chat_project/streamlit/requirements.txt
kafka-python==2.0.2
streamlit==1.29.0
streamlit_server_state==0.17.1
pandas==1.5.2
